requirements.txt
anthropic>=0.40.0
python-telegram-bot>=21.0
apscheduler>=3.10.4
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
python-dotenv>=1.0.0
aiosqlite>=0.20.0
feedparser>=6.0.11
reportlab>=4.0.0
python-docx>=1.1.0
pydantic>=2.0.0
groq
fastapi>=0.110.0
uvicorn>=0.29.0
